v2 - commit 2
fixed equal hitTime detection issue. The if statement to check if there's interupt is placed after collision check to allow both players to set hurt on other player before the if (this.isHurt) exits the while loop.

hitTime also set to undefined after each player are hurt. To make sure states are reset after a hit